DACT is now 64-bit filesize friendly.. I hope.
I changed the CRC algorithm to Adler32, and added a check for old versions
of DACT to use the old CRC algorithm.

I renamed stdint.h to dact_stdint.h (though this file isn't used anymore?)

I changed the old CRC algorithm function from `ELFCRC' to `elfcrc'

FossilOrigin-Name: fa46591195054e96538a1c4fc575069bbd01233f

+++ dact_common.h
@@ -6,5 +6,5 @@ int dact_config_execute(const char *cmd, char *options, uint32_t *blksize);
 void dact_config_loadfile(const char *path, char *options, uint32_t *blksize);
 uint32_t dact_blk_decompress(char *ret, const char *srcbuf, const uint32_t size, const char *options, const int algo, uint32_t bufsize);
 uint32_t dact_blk_compress(char *algo, char *ret, const char *srcbuf, const uint32_t size, const char *options, uint32_t bufsize);
-uint32_t dact_process_file(const int src, const int dest, const int mode, const char *options, const char *filename, uint32_t *crcs, uint32_t dact_blksize, int cipher);
+uint64_t dact_process_file(const int src, const int dest, const int mode, const char *options, const char *filename, uint32_t *crcs, uint32_t dact_blksize, int cipher);
 #endif
